requirements.txt
Flask
Flask-Login
Flask-RESTful
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
mysql-connector-python
PyMySQL
mysqlclient
cryptography
WTForms
pytz
Werkzeug
requests
stripe
gunicorn
sqlalchemy 